% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_package_deps.R
\name{all_remote_package_deps}
\alias{all_remote_package_deps}
\alias{remote_package_deps}
\alias{missing_remote_deps}
\alias{install_missing_remote_deps}
\title{Get Remote Package Dependencies}
\usage{
all_remote_package_deps(path = "DESCRIPTION", ...,
  dependencies = c("Depends", "Imports", "LinkingTo", "Suggests"),
  exclude_remotes = TRUE)

remote_package_deps(..., dependencies = c("Depends", "Imports", "LinkingTo",
  "Suggests"))

missing_remote_deps(path = "DESCRIPTION", ...)

install_missing_remote_deps(path = "DESCRIPTION", ...)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{...}{arguments passed to \code{\link{get_remote_package_dcf}}}

\item{dependencies}{List of dependencies to parse, passed to
\code{\link{get_dep_table}}}
}
\value{
A \code{data.frame} of Dependencies
(\code{all_remote_package_deps}) or a list of a
package dependency table and a vector of package names for the
remotes (\code{remote_package_deps})
}
\description{
Get Remote Package Dependencies
}
\examples{
remote_package_deps("stnava/ANTsRCore")
path = example_description_file()
all_remote_package_deps(path = path)
missing_remote_deps(path = path)
}
