% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deployed_tarballs.R
\name{deployed_tarball_version}
\alias{deployed_tarball_version}
\alias{deployed_tarball}
\title{Deployed Tarball Version}
\usage{
deployed_tarball_version(repo = "stnava/ITKR", tag = "latest")

deployed_tarball(repo = "stnava/ITKR", tag = "latest", version = NULL)
}
\arguments{
\item{repo}{GitHub repository name}

\item{tag}{Tag for the tarball}

\item{version}{version of package tarball if applicable}
}
\value{
Version number character vector
}
\description{
Downloads a source tarball from GitHub and parses the
version in that source
}
\examples{
deployed_tarball_version(  repo = "stnava/ITKR", tag = "latest")
}
