% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_remotes.R
\name{drop_remotes}
\alias{drop_remotes}
\title{Drop Remotes from DESCRIPTION}
\usage{
drop_remotes(path = "DESCRIPTION", drop_remotes = NULL)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{drop_remotes}{Packages for Remotes to Drop}
}
\value{
Rewrites the DESCRIPTION file without remotes
}
\description{

}
\examples{
path = example_description_file()
x = drop_remotes(path)
readLines(x)
}
