% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_description_file.R
\name{example_description_file}
\alias{example_description_file}
\title{Example DESCRIPTION File}
\usage{
example_description_file()
}
\value{
Path of DESCRIPTION file
}
\description{
Copies example DESCRIPTION file for parsing
}
\examples{
example_description_file()
}
