% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dep_table.R
\name{get_dep_table}
\alias{get_dep_table}
\title{Get Dependency Table}
\usage{
get_dep_table(path = "DESCRIPTION", dependencies = c("Depends",
  "Imports", "LinkingTo", "Suggests"), limit_compare = FALSE)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{dependencies}{List of dependencies to parse}

\item{limit_compare}{only keep rows where a comparison operator is found}
}
\value{
A \code{data.frame} of
\describe{
\item{name}{name of the package}
\item{compare}{comparison operator}
\item{version}{version of the package}
}
}
\description{
Get a Table of the Dependencies
}
\examples{
path = example_description_file()
get_dep_table(path = path)
}
