% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_package_dcf.R
\name{get_remote_package_dcf}
\alias{get_remote_package_dcf}
\alias{remote_package_dcf}
\alias{has_remote_dcf}
\title{Read Remote Package DESCRIPTION file}
\usage{
get_remote_package_dcf(
  remotes,
  pat = NULL,
  url = c("https://github.com", "https://api.github.com"),
  output_directory = NULL,
  overwrite = FALSE,
  ...
)

remote_package_dcf(...)

has_remote_dcf(remotes, pat = NULL, url = "https://github.com", ...)
}
\arguments{
\item{remotes}{Repository name}

\item{pat}{GitHub Personal Authentication Token (PAT)}

\item{url}{URL for repo - github}

\item{output_directory}{Output directory for the description file to download to,}

\item{overwrite}{overwrite the file if exists.  Only really relevant if
\code{output_directory} is not \code{NULL}}

\item{...}{additional arguments to pass to \code{\link{GET}}}
}
\value{
List of Package/Version DCF information
}
\description{
Read Remote Package DESCRIPTION file
}
\examples{
get_remote_package_dcf("stnava/ANTsRCore")
out = get_remote_package_dcf("stnava/ANTsRCore", output_directory = tempdir())
out = get_remote_package_dcf("stnava/ANTsRCore", output_directory = tempdir())
remote_package_dcf("stnava/ANTsRCore")
remote_package_dcf("stnava/ANTsRCore", output_directory = tempdir())
}
