% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_remotes.R
\name{get_remotes}
\alias{get_remotes}
\title{Get the Remotes from a Package}
\usage{
get_remotes(path = "DESCRIPTION")
}
\arguments{
\item{path}{Path to DESCRIPTION file}
}
\value{
Character vector of remotes
}
\description{
Parse a DESCRIPTION file and returns the remotes
}
\examples{
path = example_description_file()
get_remotes(path)
}
