% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_remote_binaries.R
\name{install_remote_binaries}
\alias{install_remote_binaries}
\title{Install Binaries from Remotes}
\usage{
install_remote_binaries(path = "DESCRIPTION", package = NULL,
  drop_all = TRUE)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{package}{Specific package to install}

\item{drop_all}{If TRUE, then will only drop packages that are installed}
}
\value{
Invisible NULL
}
\description{
Gets Remotes, gets Binaries form GH, and installs them, then
drops them from Remotes field
}
\examples{
\dontrun{
path = example_description_file()
install_remote_binaries(path = path)
}
}
