% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_remotes_no_dep.R
\name{install_remotes_no_dep}
\alias{install_remotes_no_dep}
\title{Returns the specific package from Remotes}
\usage{
install_remotes_no_dep(path = "DESCRIPTION", package = NULL,
  reorder = TRUE, drop = TRUE, verbose = TRUE, force = FALSE, ...)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{package}{Package to subset.  If NULL, then all will be installed}

\item{reorder}{should remotes be reordered before running}

\item{drop}{Remotes should be dropped after installing}

\item{verbose}{Print diagnostic message}

\item{force}{should the command run even if install_github fails?}

\item{...}{arguments to pass to install_github}
}
\value{
Character vector of remotes

Output of \code{\link{install_github}}, indicator
of TRUE/FALSE for installation
}
\description{
Subsets the remotes based on a package
}
\examples{
\dontrun{
  install_remotes_no_dep()
}
}
