% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_release_with_binary.R
\name{latest_release_with_binary}
\alias{latest_release_with_binary}
\title{Get the Latest Binary for a repository}
\usage{
latest_release_with_binary(repo, pat = NULL, verbose = TRUE,
  check_r_version = TRUE, force_sha = TRUE, ...)
}
\arguments{
\item{repo}{Remote repository name}

\item{pat}{GitHub Personal Authentication Token (PAT)}

\item{verbose}{print diagnostic messages}

\item{check_r_version}{Check if R version is in the name of the
tarball}

\item{force_sha}{Should the SHA be checked when trying to get
binaries}

\item{...}{additional arguments to \code{\link[httr]{GET}}}
}
\value{
URL of the binary
}
\description{
Parses a remote and gets a URL for a binary release
}
\examples{
repo = "stnava/ANTsR"
res = latest_release_with_binary(repo)
res
repo = "neuroconductor/gifti"
res = latest_release_with_binary(repo)
res #'
}
