% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_these_packages.R
\name{package_comparison_table}
\alias{package_comparison_table}
\alias{update_these_packages}
\title{Get Comparison Table from Dependencies}
\usage{
package_comparison_table(path = "DESCRIPTION", dependencies = c("Depends",
  "Imports", "LinkingTo"))

update_these_packages(...)
}
\arguments{
\item{path}{Path to DESCRIPTION file
passed to \code{\link{remote_binaries}}}

\item{dependencies}{List of dependencies to parse}

\item{...}{not used}
}
\value{
Invisible NULL
}
\description{
Gets Remotes and compares to installed packages
}
\examples{
path = example_description_file()
update_these_packages(path = path)
}
