% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_name.R
\name{package_name}
\alias{package_name}
\title{Get Package name}
\usage{
package_name(path = "DESCRIPTION")
}
\arguments{
\item{path}{Path to DESCRIPTION file}
}
\value{
Name of Packge
}
\description{
Get Package name
}
\examples{
path = example_description_file()
package_name(path)
}
