% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_needs_update.R
\name{package_needs_update}
\alias{package_needs_update}
\title{Check Package Versions}
\usage{
package_needs_update(version, package)
}
\arguments{
\item{version}{Version of package}

\item{package}{Package to check}

\item{...}{not used}
}
\value{
Output from \code{\link{compareVersion}}
}
\description{
Looks in installed packages to see if a new version is needed
}
