% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_one_remote.R
\name{parse_one_remote}
\alias{parse_one_remote}
\title{Parse a Remote file}
\usage{
parse_one_remote(x)
}
\arguments{
\item{x}{Parses a remote field}
}
\value{
A repo object, from \code{devtools}
}
\description{
Parse Remote from DESCRIPTION
}
\examples{
parse_one_remote("stnava/ANTsR")
}
