% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_version.R
\name{r_version}
\alias{r_version}
\title{Get R Version}
\usage{
r_version()
}
\value{
A character string with just major and
minor (no patch or hotfix) version of R
}
\description{
Get R Version
}
\examples{
r_version()
}
