% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_binaries.R
\name{remote_binaries}
\alias{remote_binaries}
\title{Get Binaries from Remotes}
\usage{
remote_binaries(remotes = NULL, path = "DESCRIPTION", package = NULL)
}
\arguments{
\item{remotes}{Remotes to get the binaries for}

\item{path}{Path to DESCRIPTION field}
}
\value{
List of remote binaries
}
\description{
Gets Remotes, gets Binaries form GH
}
\examples{
remote_binaries(remotes = "muschellij2/neurobase")
path = example_description_file()
remote_binaries(path)
}
