% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_order.R
\name{remote_order}
\alias{remote_order}
\alias{reorder_remotes}
\title{Order of the Remotes for Installation}
\usage{
remote_order(path = "DESCRIPTION", ..., max_iter = 200)

reorder_remotes(path = "DESCRIPTION", verbose = TRUE, ...)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{...}{arguments passed to \code{\link{remote_package_deps}}}

\item{max_iter}{Number of iterations - shouldn't have to change,
stops from infinite loop}

\item{verbose}{Print diagnostic message}
}
\value{
Order of remotes
}
\description{
Order of the Remotes for Installation
}
\examples{
path = example_description_file()
x = remote_order(path = path)
x
}
