% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_remote.R
\name{subset_remote}
\alias{subset_remote}
\title{Returns the specific package from Remotes}
\usage{
subset_remote(path = "DESCRIPTION", package = "ANTsR")
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{package}{Package to subset}
}
\value{
Character vector of remotes
}
\description{
Subsets the remotes based on a package
}
\examples{
path = example_description_file()
desc = get_remote_info(path)
}
