% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_table.R
\name{tag_table}
\alias{tag_table}
\title{Tag Table}
\usage{
tag_table(repo, pat = NULL, mustWork = TRUE, ...)
}
\arguments{
\item{repo}{Remote repository name}

\item{pat}{GitHub Personal Authentication Token (PAT)}

\item{mustWork}{Does the command need to work?  If fails,
will error.  But \code{FALSE} will try to pass
appropriate missing data value}

\item{...}{additional arguments to \code{\link[httr]{GET}}}
}
\value{
\code{data.frame} of tags
}
\description{
Parses a remote and gets a table of tags
}
\examples{
repo = "stnava/ANTsR"
tag_table(repo)
tag_table("muschellij2/ANTsR")
tag_table("cran/psych@084bdd0ae2630cf31c26d97a6e13e59d3f0f66e6")
}
