% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_release_table.R
\name{binary_release_table}
\alias{binary_release_table}
\alias{binary_table_no_tags}
\title{Get Binary Release Table for a repository}
\usage{
binary_release_table(repo, pat = NULL, ...)

binary_table_no_tags(repo, pat = NULL, ...)
}
\arguments{
\item{repo}{Remote repository name}

\item{pat}{GitHub Personal Authentication Token (PAT)}

\item{...}{additional arguments to \code{\link[httr]{GET}}}
}
\value{
\code{data.frame} of binary releases
}
\description{
Parses a remote and gets a table of binary releases
}
\examples{
repo = "stnava/ANTsR"
binary_release_table(repo)
binary_release_table("muschellij2/ANTsR")
repo = "stnava/ANTsR"
binary_table_no_tags(repo)
binary_table_no_tags("muschellij2/ANTsR")
}
