% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_desc_elements.R
\name{cat_desc_elements}
\alias{cat_desc_elements}
\alias{cat_desc_elements_remote}
\title{Print Description Elements out}
\usage{
cat_desc_elements(path = "DESCRIPTION")

cat_desc_elements_remote(repo, ...)
}
\arguments{
\item{path}{path to DESCRIPTION file}

\item{repo}{remote repository on GitHub}

\item{...}{Additional arguments passed to \code{\link{get_remote_package_dcf}}}
}
\value{
Nothing
}
\description{
Print Description Elements out
}
