% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote_package_dcf.R
\name{get_remote_package_dcf}
\alias{get_remote_package_dcf}
\alias{remote_package_dcf}
\alias{has_remote_dcf}
\title{Read Remote Package DESCRIPTION file}
\usage{
get_remote_package_dcf(remotes, pat = NULL, url = "https://github.com", ...)

remote_package_dcf(...)

has_remote_dcf(remotes, pat = NULL, url = "https://github.com", ...)
}
\arguments{
\item{remotes}{Repository name}

\item{pat}{GitHub Personal Authentication Token (PAT)}

\item{url}{URL for repo - github}

\item{...}{additional arguments to pass to \code{\link{GET}}}
}
\value{
List of Package/Version DCF information
}
\description{
Read Remote Package DESCRIPTION file
}
\examples{
get_remote_package_dcf("stnava/ANTsRCore")
remote_package_dcf("stnava/ANTsRCore")
}
