% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_remote_binaries.R
\name{install_remote_binaries}
\alias{install_remote_binaries}
\title{Install Binaries from Remotes}
\usage{
install_remote_binaries(path = "DESCRIPTION", remotes = NULL,
  package = NULL, check_r_version = FALSE, verbose = TRUE,
  method = "wget", ...)
}
\arguments{
\item{path}{Path to DESCRIPTION file
passed to \code{\link{remote_binaries}}}

\item{remotes}{Remotes to get binaries for - in case not going from DESCRIPTION,
passed to \code{\link{remote_binaries}}}

\item{package}{Specific package to install}

\item{check_r_version}{Check if R version is in the name of the
tarball}

\item{verbose}{Print diagnostic messages}

\item{method}{method to download, passed to \code{\link{download.file}}}

\item{...}{additional arguments passed to \code{\link{install.packages}}}
}
\value{
Invisible NULL
}
\description{
Gets Remotes, gets Binaries form GH, and installs them, then
drops them from Remotes field
}
\examples{
\dontrun{
path = example_description_file()
install_remote_binaries(path = path)
}
}
