% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_remotes_no_dep.R
\name{install_remotes_no_dep}
\alias{install_remotes_no_dep}
\title{Returns the specific package from Remotes}
\usage{
install_remotes_no_dep(path = "DESCRIPTION", package = NULL,
  reorder = TRUE, drop = TRUE, ...)
}
\arguments{
\item{path}{Path to DESCRIPTION file}

\item{package}{Package to subset.  If NULL, then all will be installed}

\item{reorder}{should remotes be reordered before running}

\item{drop}{Remotes should be dropped after installing}

\item{...}{arguments to pass to install_github}
}
\value{
Character vector of remotes

Output of \code{install_github}
}
\description{
Subsets the remotes based on a package
}
\examples{
\dontrun{
  install_remotes_no_dep()
}
}
