% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest_release_with_binary.R
\name{latest_release_with_binary}
\alias{latest_release_with_binary}
\title{Get the Latest Binary for a repository}
\usage{
latest_release_with_binary(repo, pat = NULL, ...)
}
\arguments{
\item{repo}{Remote repository name}

\item{pat}{GitHub Personal Authentication Token (PAT)}

\item{...}{additional arguments to \code{\link[httr]{GET}}}
}
\value{
URL of the binary
}
\description{
Parses a remote and gets a binary Get the Latest Binary for a repository
}
\examples{
\dontrun{
repo = "stnava/ANTsR"
latest_release_with_binary(repo)
}
}
