% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_table.R
\name{tag_table}
\alias{tag_table}
\title{Tag Table}
\usage{
tag_table(repo, pat = NULL, ...)
}
\arguments{
\item{repo}{Remote repository name}

\item{pat}{GitHub Personal Authentication Token (PAT)}

\item{...}{additional arguments to \code{\link[httr]{GET}}}
}
\value{
\code{data.frame} of tags
}
\description{
Parses a remote and gets a table of tags
}
\examples{
repo = "stnava/ANTsR"
tag_table(repo)
tag_table("muschellij2/ANTsR")
}
